gracefully close connection
first go at protocol for gracefully closing a connection.
This implementation does not yet make best effort to wait for the send
buffer to drain if the connection is closed with pending messages in the
send buffer. If a message is taken from the send buffer which is aready
closed / being closed, the publisher is notified that the message cannot
be sent. The publisher needs to account for this to happen anyway but
the implementatation could be more friendly in these siztations...

related to #2
closes #3