feat(index): add block buffer cache
An LRU cache allowing the hash index to load a maximum nunber of blocks from
the index-store. (Blocks contain the entries managed by the hashindex).

* buffer writes: if multiple entries are written to the same block, the
  block does not need to be written back to the index store on every
  write but only when it is evicted from the cache. This effectively
  "buffers" the writes and batches them up into a single write
  operation.
* cache reads: if an entry needs to be read from a block and that block
  is currently located in the cache, the block does not need to be
  loaded from index store, saving the load operation.

Adds a HashIndex.flush() method which can be called to flush the cache
and force all buffers to be written back to the index store. This method
must be called before operations like snapshots which directly access
the index store.

related to #7