fix(async-start): behavior on concurrent async start / remove
Fixes and formalizes behavior in this situation where remove() is
performed concurrently to async start.

The behavior is as follows:

* when remove is called, the operation of the service is set from
INSTALLING to REMOVING
* first, the container waits for the async start to complete. Then,
depending on whether the start completed sucessfully or not, the
service is stopped.

In detail, this means:

* if the async start completed successfully, then, stop() is invoked
on the service so that it can close any resources it holds. Then
it continues with the regular stop sequence.
* if the async start is completed exceptionally, then stop() is not
invoked.

related to #7