fix(Filter): do not display placeholder as value in IE11
Internet Explorer does not play nice with the combination of Placeholders
and the Attribute component that sets the type. In this combination, the
placeholder text is set as value for the input.

To fix this, we use the hidden css class instead of setting the input type

related to OPT-460