chore(broker-core): increase spin/yield iteration of BackoffIdleStrategy
If the broker is run on a bigger system with more then 2 threads we saw
throughput issues. As there a more threads to use for the actor scheduler
threads where often triggering the idle strategy which will reduce the
throughput. This commit increases the spin/yield counts before the
backoff idle strategy actually starts to sleep.

Additionally the performance test warmup phase will run with a higher
request rate to not trigger the idle strategy before the actual test phase
starts.