chore(state-machine): count transition as work count
* Increase work count if current state has changed. This has effect for
actors which has a state machine with wait states. If a wait state is
left (e.g. because a timeout is reached or a response is available) then
the actor may continue with the following state. Previously, the actor
had to wait for the next work cycle (because the work count was zero).