feat(metrics): task execution metrics
For each task the following metrics are recorded:

- execution count: how often the task was executed
- total execution time
- max execution time

Tradeoff 1:

Metrics can be enabled on a per task basis. This is a tradeoff between
the overhead caused by allocating the counters in the shared buffer vs.
efficiency of scheduling. Short-lived tasks should be very cheap to
schedule and probably do not need these metrics. When sheduling
long-lived tasks, a small additional overhead when initially submitting
the task to the scheduler is acceptable.

Tradeoff 2:

Note that task execution times will include scheduler overhead. The goal
is that eventually, the scheduler overhead will be more or less constant
(ie. O(1) and wait free)

related to zeebe-io/zeebe#671