feat(snapshotReplication): expose snapshot properties for listing in broker
getName() is admittedly perhaps a misnommer, and should be storageId() or
some such, since there could be collisions where a name refers to more than
one stored snapshot (only an issue at the moment when a broker would be
downloading a snapshot while the leader would overwrite it). For a first
iteration this is fine, but it should be handled later.

related to zeebe-io/zeebe#876