chore(client): fix race condition when closing topic subscription
closing a SubscriberGroup is done while blocking the current phase,
but the future it relies on gets resolved in jobs that do not block
the current phase. this means there is a race condition, where if you
close the actor and all jobs get discarded (but the ones blocking
in current phase), then the future never gets resolved.

related to zeebe-io/zeebe#939