feat(snapshotReplication): add test case for interleaved requests
adds test case for when requests are interleaved with responding
for the requests (e.g. req A comes in, schedules response A',
req B comes in, schedules response B', response A' is sent out,
response B' is sent out).

useful since we reuse objects to minimize allocation for snapshot
replication requests.

related to zeebe-io/zeebe#878