chore(broker): fix potential endless loop in Management API
if the response supplier throws an exception, the actor could be stuck
in an endless loop since neither yield nor done is ever called.

current strategy is thus to mark as done on exception, though in the
future maybe a yield with retry limit would be better.

related to zeebe-io/zeebe#878