fix(scheduler): handle race condition where job is added to dropped queue
If the submit method is interrupted between getting the reference of the
submitted jobs queue and the offer of the new job. And in the meantime the
onClosed method is invoked and drains the old queue before the offer happens
the job will still be added to the original queue and never executed.

To prevent this we check after the offering of the job if the submitted
jobs queue was replaced in the meantime and fail the job manually. We ignore
exceptions in case the job is failed multiple times, i.e. during the drain
and after the swap check.