ci(surefire):  rerun flaky tests up to 3 times before failing the build
The test suite has some flaky tests which reduces the value of the CI as the
developer has to analyse the build log and decide whether a failed test is a
real problem or flaky test. To reduce this we will now rerun failing tests
up to 3 times before failing the build.