fix(BPMNDiagram): fix weird race condition
Sometimes a diagram gets unmounted shortly after its xml property
changed. In that case, there might be an asynchronizity between the xml
and the viewer model. To fix that, we make sure that the unmount
lifecycle method has no access to the outdated viewer instance by
dereferenzing it when the viewer is detached from the dom

related to OPT-1331