chore(gateway): allow runOnCompletion with ResponseFuture
This change allows the gateway to send requests without using join. Instead the
gateway can now await completion of response futures using the actor scheduler.
This better fits the async observer pattern of the gRPC server interface.

The problem was that the request manager internally waited blocking on an
ActorFuture if the topic was not known in the topology. A blocking get call on
an ActorFuture inside the actor scheduler is not allowed. This change now
assumes that the partition id lookup might be async and defers the actual send
of the request until the partition is resolved.

Changes to migrate the gateway to facilitate this are postponed to later
commits.