feat(broker): add debug exporter
Adds an exporter implementation for debugging. It will log every
exported record using the logger provided to the exporter. It can be
configured to use a specific log level (default: debug) and if records
should be pretty printed for better readability.

The debugger can be enabled for every embedded broker/clustering rule
during testing by setting the ENABLE_DEBUG_EXPORTER to true.

The default distro configuration contains a exporter example for the
debug exporter so users can easily enable it.

The debug exporter does not update the exported records position by
intention so every record is always printed even on restart of the
broker.