fix(alerts): use correct link in email body
* Also stabilized it reminder tests. When there was more time
passing than one second between schedule of the reminder and
and the manual trigger of the reminder job, it could not find
the reminder job anymored, since the job has already been triggered
before. Hence, he was removed before the manual trigger and caused
a test failure.

related to #OPT-1587