fix(broker): ensure YAML transformation does not break the msgpack serialization
This does not fix the underlying problem of the msgpack implementation but
is a quickfix as the YAML transformation seems to be the only place where we
change the content of memory view into a msgpack buffer while iterating over
it.

To solve the underlying problem we would have to create a memory copy of the
modified record before updating the iterator buffer. This would add some
overhead which is at the moment not required.