fix(client/Authentication): send initial request to check login status
Instead of passively waiting for an unauthorized response the Authentication
component now sends an initial request to the user API. Until it receives the
response an empty page will be rendered. Based on the response a redirect to
the login page or the normal page will be shown.

Related to OPE-737