fix(elasticsearch-exporter): ignore variable values above 32766 bytes
The type keyword in elasticsearch has a default max length of 32766 bytes.
A large variable value can exceed that limit. If we don't ignore these large
values the es exporter fails and loops on exporting the variable value.

By ignoring the value it is exporter and stored in the _source attribute in
elasticsearch but not index anymore.

For more information see https://www.elastic.co/guide/en/elasticsearch/reference/6.8/ignore-above.html