chore(benchmarks): add more starters per default
To improve our test coverage I added a simple starter and a timer starter. The simple one contains just
start event and end event. It is quite interesting to see how the latency behaves on these instances.
The timer starter contains a workflow with a timer set to PT1S. Per default we will start all of them
to have high load and test that this can be handled by the brokers.