chore(gateway): add flag to enable/disable long polling
This is a combination of 7 commits (incl the 1 above):

chore(gateway): enable long polling by default

chore(gateway): improve longpolling flag name

chore(gateway): test activate jobs with/without long polling

chore(gateway): clarify naming and add documentation

chore(gateway): add defaults to config templates

test(gateway): add IT to test job activation without long polling

In order to test for this, we need to set
zeebe.broker.gateway.longPolling.enabled to false. This configuration
can only be changed at start-up of the broker and so this could not be
tested in the same class as ActivateJobsTest, which was testing
activatejobs behavior for the default long polling enabled.

Since a LongPollingDisabledActivateJobsTest seemed like a strange test,
I've renamed ActivateJobsTest to LongPollingActivateJobsTest and created
a new ActivateJobsTest that tests it with long polling explicitly
disabled.