chore(broker): improve streamprocessor time safety
The healthcheck tick should be performed before the open future is
completed, to make sure it is initialized when the health monitoring
starts.

This commit also changes how failures are handled. Before this commit
they were handled internally in the class and rethrown. The unhandled
thrown exception was then handled by the Actor which in turn calls the
handleFailure method. This lead to an error message where the actor
failed in phase FAILED, while it actually failed in phase STARTED and
then the actor was failed afterwhich the failure was handled again.