docs(forks): add concurrency considerations
This commit adds additional descriptions on how to deal with parallel
flows in workflows, considering race conditions. These are especially
important when dealing with call-activities, because they merge all
variables the child instance contains into the parent instance (but the
child also starts with all variables the parent has, so if the parent
makes changes this is a race condition), so always make sure to use
output mapping on a parallelised call activity.