fix(atomix): use ip to map to channel pool
When the broker with an address is restarted, it will have same address but different ip. When address is used to map to the channel, this can lead to the sender sending messages to the channel connected to the old ip instead of creating a new channel.

(cherry picked from commit cb0a7330d6b301f3403eca6bdc5ec4565e5ab394)