test(engine): fix flaky CancelWorkflowInstanceConcurrentlyTest
There was a timing issue in the multi-instance scenario, where the test
did not wait for both of the jobs to be created. Then when re-starting
the engine, reprocessing occurs including the verification of
reprocessing which would then fail because of this inconsistency.

By waiting for 2 jobs to be created and then taking the 1 of the correct
type, instead of just waiting for the 1 job, the test has become stable..

Co-authored-by: Philipp Ossler <philipp.ossler@gmail.com>