test(engine): fix flaky test again
The test has multiple scenarios, and the sequential scenario actually
only expects 1 job to become activatable.

In addition, in the scenarios where we expect 2 jobs to become
activatable, it was not enough to just limit the stream to 2 when only
grabbing one specific one. This because it might not be clear which job
is created first. Therefore, the tests should always first assert that
all expected jobs have been created and then continue.