fix(borker): fix concurrent partition closing
Allows to close and transition inorder. Previous multiple transitions
and closing have worked against each other, which means it can lead to
unexepected out comes or dead locks.

Transitions are now executed in a sequential order. The closing will
await the current transition, before starting the next transition and
closing all resources.

Add new tests for transition from different Role and to close the
partition when currently an transition is on going