fix(clients/java): improve job worker scheduling
This commits fixes 2 ways in which the scheduling to poll for jobs could
stop in the JobWorkerImpl. The main change is the using the correct
value to determine whether or not to reschedule onPollSuccess. In the
time that the onPollSuccess is called with some number of activated
jobs, it can be that all these jobs have already been handled
completely. In that case, the onPollSuccess should reschedule, because
the number of remainingJobs has dropped to 0.

The second improvement is to reschedule in the case the job worker
manages to reach the inner critical section of the polling, while
actually not yet needing to poll (not having dropped below the
threshold). It might not always hold that another scheduled poll already
exists, so it makes sense to schedule a new one.