chore(engine): transition to initial state
Migrated bpmn element processors will implement the processing of the
`ActivateElement`, `CompleteElement` and `TerminateElement` commands.

To provide a good context inside the processors, this commit makes it so
that the command processing automatically transitions the state to the
initialising part of the states (activating for activate, completing for
complete and terminating for terminate).