refactor(test-util): support child process generation
In order to support call activities, the random process generator should
be able to provide all generated processes (for the main process, as
well as the child processes) for deployment.

The ProcessBuilder now passes a callback function to the
constructioncontext, to allow other builders to inform the root process
of any constructed child processes.

Apart from the naming scheme of bpmn files written to disk when running
`RandomProcessGenerator.main(String[] args)`, no behavioural changes are
achieved in this commit. I've thus marked it as refactor.