feat(protocol): new ProcessEvent value type
Introduces a new ProcessEvent value type and intent. The new value is
meant to reflect event triggers emitted by flow nodes targetting
receivers that are not directly related to them, and for which there are
no special purpose entities.

This came out because of the error throw end event, which may be caught
by an element that is not directly the flow scope of the element, and
for which there is no related entity (subscription, timer, etc.) which
could be triggered, thus making the change "untraceable" via events.

By adding a new record, we make the change to the state explicit for
such things. This can later be reused for escalation, compensation, and
possibly other such events.