ci(github): cut off pr body in bors merge commit
This commit adds a specific marker to cut off the merge commit body.

The definition of ready and definition of done checklists add a lot of
unnecessary volume to merge commits, because bors copies the pr
description into the merge commit body. We already instructed bors to
cut off these bodies at a marker, but the marker has been removed at
some point. Having a specific marker for it reduces the change we remove
it by accident.