chore(engine): add value metric about root process instances
The new metric "zeebe_executed_instances_total" counts executed root
process instances, and can be extend for other use cases. The metric
exposes a organization label which can be used to aggregate these values
over an organization in Camunda Cloud. The organization is set to "null"
in case the env variable "CAMUNDA_CLOUD_ORGANIZATION_ID" is not set.