chore(atomix): readers always start at first index after creation
There is no value in specifying an index when creating a reader because always the reader is started at first index or last index. It is cleaner to make a reader always starts at the first index. The creater of the reader can then seek to the specific index if needed.