chore(engine): split up onElementCompleted in 2 methods
The problem with the code was that - for migrated element processors - this method was called after
the element had already completed.
Therefore, the element and all its variables were already removed from storage.
The change introduces two methods before/after completed of the element. This allows to separate steps
that need to happen before the element is deleted from the steps that can happen after the element is deleted.

The method was also renamed from onChildCompleted which sournds like it should be called whenever an
element was completed - which was not the case - to before/AfterExecutionPathCompleted which better
reflects the semantics of the method.