fix(engine): set inner instance loop counter correctly
Update the loop counter on the element instance object
(flowscopeInstance in this code) when incrementing the loop counter in
the state.

This is necessary because: When a migrated element type is the inner
instance of a multi instance body, the loop counter on the inner
instance is set to the unincremented value, i.e. the loop counter is
incremented in the state on the multi-instance body, but then the loop
counter value is taken from the multi-instance body element instance
that is kept in memory which was not modified by the increment method.