refactor(snapshot): ignore checksum files on load
As the checksum file is now a sibling file of the snapshot, it gets
picked up during startup as a potential snapshot itself, resulting in
weird warnings that it's not a valid snapshot.

The simplest fix is to ignore checksum files when loading the latest
snapshot, as we expect that file to be present.