refactor(engine): extract reprocessing workaround
The BpmnStreamProcessor contains some code that deals specifically with
unmigrated element records, because they can lead to reprocessing
problems. In the near future, this code is expected to grow a bit more
before it can be removed. This commit extracts it into it's own method
and applies early returns to ease complexity when reading.