feat(engien): migrate completion and termination of process
Migrate to event sourcing the completion and termination of a process.

If process was created by a call activity we need to look up the related
element instance before terminating/completing the process element
instance. Afterwards we can terminate/complete the parent (call
activity), to keep the order of events.

ProcessInstanceElementCompletedApplier does propagation of variables to
parent call activity if exist, such that no state change is done in the
ProcessProcessor.