chore(journal): use frame header to check if next entry exists
* hasNext uses the frame to see if there is a next entry so that it doesn't have to read the nextEntry.
We had to reset all readers when truncating because the in-memory state (nextEntry) can be corrupted. If we don't keep any in-memory state, only the readers which are at the truncated entries have to be reset. This helps us to alleviate some of the concurrency issues between truncate and readers.

* Ensure reader util never returns null. If the frame have a valid version, then we expect to read a valid record. If we cannot it should be considered as corruption.