chore(test-util): logical predecessor for parallel gateways
In parallel gateways two or more branches can be executed in parallel. The execution path however, defines
a strict serialization of the steps on all tasks. This leads to situation where several BPMN elements are
activated at once and waiting to be executed. This is where we need to keep track of the logical predecessor
for each step. Furthermore, this needs to work recursively.