chore(test-util): add boundary timer event to sub process execution
This adds boundary timer events to the execution path for subprocesses.
These timers have one of two timeouts:
- If they are not part of the execution path, they are initialized to VIRTUALLY_INFINITE
- If the are part of the execution path then the actual timeout is calculated for them

This is achieved by adding a method to AbstractExecutionStep that allows to update the variables
for a given execution path and the activation time that is calculated based on that.

Also introduces a new interface for steps that can start a process. This is to better distinguish between
the general variables of an execution step and the calculated variables to start a process.