fix(journal): make truncate and delete safe with concurrent readers
Use StampedLock for concurrency control between deleteAfter/reset operation by a writer and all operations by readers. It is not necessary for append to acquire lock, as it is doesnot cause an inconsistent state observed by the reader. As a result, this is read heavy workload because deleteUntil and reset are only called infrequently.