chore(engine): refactor mechanism to cut off execution path
This became necessary because the old mechanism was flawed: It did check, wherther an execution path can be interrupted, which is good.
However, then it used a naive random index for the cutoff point. This was insufficient because an execution path can only be interrupted
at certain points. And the random index might pick a position at which the execution path cannot be interrupted.

The new algorithm is more sophisticated to find the cutoff index. Also, the logic moved to ExecutionPathSegment. The hope is that this logic
will be reused for other blocks which have boundary events.

Finally, two methods in ExecutionPathSegment were deprecated. Currently, they are never called. The only code that references them is in
ProcessBuilder, but it is dormant. Fixing those would have increased the scope of this change. Plus, in my opinion, the dormant code needs
more rework anyway.