fix(engine): set temp vars for event sub process
The start event of an event sub process wants to apply output mappings
for that the temp vars need to exist on the correct scope (flow scope),
which is the event sub process. In order to make that possible the temp
vars need to be copied from the flow scope of the event sub process into
the scope of the event sub process. The
`ProcessSubscriptionMessageCorrelatedApplier` currently sets these
variables on the flow scope of the triggered event sub process.

Tmp vars are now copied into the event sub process on input mapping
execution. Later these should be migrated to the EventApplier.

Co-authored-by: Nico Korthout <nico.korthout@camunda.com>