feat(engine): apply event sourcing to sequence flows
Since the only record that is processed for sequence flows is an event,
simply remove the processor and move its logic else where.

The sequence flow processor kept track of taken incoming sequence flows
of a joining parallel gateway. This gateway should only be activated
once all incoming sequence flows have been taken. Keeping track of which
flows are taken is now tracked by the sequence flow taken applier.

Checking whether or not to activate the parallel gateway is a done in
the state transition behaviour.