chore(journal): don't seek if the reader is already at the index
Previously, this check was done in RaftLogReader. But after a truncate or reset oepration, the cached nextIndex in an uncommitted raftlog reader is no more consistent with the journal reader. By moving this check from raftlog to the journal, the in-memory state of the raftlog reader is not corrupted after a reset or truncate.