fix(test-util): fix IllegalArgumentException
Fixes an IllegalArgumentException when finding a cut off point for a
path segment which could occur if there was only one step in the
segment, resulting passing 0 to the random number generator as a
positive bound.

If there is only on element in the path, then we just return an empty
segment such that it can be immediately interrupted.