fix(engine): stop child processing on incident
By Processing a ACTIVATE command we will normally create a new element
instance (and go to the ACTIVIATING state for that instance), we notify
the parent execution for these activation. If this notifaction fails we
want to stop the further child execution and create for that an
incident.

Example: Input mapping of the event sub process are evaluated as part of
the start event activation. If this fails we want to create an incident.

This commit modifies the onChildActivating method, such that it returns an
Either<Failure, Void> which allows us to make the disinction between
Success and Failure on parent notifcation.

Co-authored-by: Christopher Zell <zelldon91@googlemail.com>
Co-authored-by: Nico Korthout <nico.korthout@camunda.com>