test(snapshot): ensure ActorScheduler is always closed
Ensures all usages of ActorScheduler are always closed in test, mostly
by switching to using ActorSchedulerRule where possible, or using a
try-with-resources approach.

Ensuring that we always close all resources is important when running
long living, parallelized build processes in CI - while each scheduler
doesn't use up much, if we keep opening and forgetting to close them
everywhere, it becomes death by a thousand cuts.

The one caveat is a test in FileBasedReceivedSnapshotTest,
shouldNotRemovePendingSnapshotOnCommittingSnapshotWhenHigher, which was
modified such that a third store/actor scheduler is not necessary.