feat(engine): always write ProcessEventIntent.TRIGGERING
No matter what type of event, we always write the ProcessEvent which
creates in the applier the corresponding EventTrigger with the newly
generated key.

Previous for each event it was done differently and the event trigger
had keys which depend on the type of the event which was triggered. This
makes it hard to do the clean up, which we need for #6850.

Plus this simplifies a lot the event trigger creation (where and how).