docs(CONTRIBUTING): change commit message guidelines
The commit types were used to generate changelogs in the past. However,
that is no longer the case. The team wants to use commit types that help
to write better commit messages and better describe the intent of the
changes. This commit changes the allowed commit message guidelines to
better fit the current needs of the team.

Main changes are:
- drop the `chore` type. It was overused and did not describe the intent
- change `fix` and `feat` to allow smaller changes to match
- add a set of new types from
  [Angular](https://github.com/angular/angular/blob/master/CONTRIBUTING.md)
  to have more options to choose from
- add more types (`deps` and `style`), to match our needs
- scope is now optional and has a broader definition (name of component)
  to better fit our needs