fix(monitor): use correct processing metric query
The processing panel was not correctly calculating
the records processed per seconds, since it assumed
that events AND commands are processed, which is
no longer true. Only Commands are processed on
normal processing the events are skipped, which we
need to take into account on the calculation.

Fix some panels, such that they not showing old values.
Use instant for that, which shows data from the end of
the time range instead of the complete range.

Instant queries in dashboards https://grafana.com/docs/grafana/next/datasources/prometheus/