refactor(engine): remove getDeferredRecods()
The else case in MultiInstanceBodyProcessor was removed because this case can never happen.
This was old code and presumably from an earlier stages when multi instance was not as
mature. These days, the only way an event sub process could be triggered is if it
is part of an embedded sub process. But in this case it will be triggered by the
child and not by the multi instance body.