test(qa): use admin endpoint to verify snapshots instead of reading the directory
Since this is a qa test, it is better to use the existing apis rather than exposing the internal implementations details such as snapshot directory in the tests. Also removed the comparing checksums in the tests. Checksums are internal implementation details. For a qa test it is enough to compare snapshot ids. The other things such as checksum should be covered by unit tests.