ci(jenkins): renable nightly develop QA run
Previously, the nightly QA run was only enabled on develop. Later that
was switched to the latest stable branch.

This commit renables the nightly QA run for the develop branch to run
next to the latest stable branch.

The cron trigger is defined for both the develop as well as the stable
branches, so a cron trigger exists on both. Then when the TimerTrigger
triggered the build and the branch is either develop or the latest
stable branch, then the QA step is executed.