test(qa): remove BrokerFailedLeaderChangeTest
At the time this test was added, there was not other way to test this behavior. However, now there are unit tests for ZeebePartition that test for this. It is better to remove this test because it is relying too much on internal details to test this even though it is a qa test.