refactor(snapshot): compare written checksum to expected checksum
To simplify testing, and in preparation for #6666, we compare the
written checksum on reception to the expected one instead of reading the
checksum file. This simplifies testing when we have written the wrong
checksum, and also prepares for when the checksum file will not be
included in the snapshot and thus will not be read.