feat(test-util): add DirectoryAssert
Introduces directory specific assertions, which extend
AbstractPathAssert. There was no easy way to previously assert that a
directory contains exactly only X paths, or that it contains all of the
given paths - it could be done via a combination of different
assertions, but it's simpler and easier to read with specific
assertions.