feat(snapshot): persist checksum as sibling file
Persists the snapshot checksum as a sibling file outside of the
snapshot, and caches the value on the persisted snapshot instance
itself. This is done at persist time, and on start up when loading the
previous snapshot.

Introduces a new error, InvalidSnapshotChecksum, which is thrown at
start up if either the checksum file exists but could not be read, or
the snapshot has a mismatched checksum.

Validation of the checksum is done in the store now, after the pending
snapshot (both transient/received) has been moved to its final
destination. This is to verify the integrity of the persisted snapshot,
as the checksum file is used as a marker file to know if the persist
operation was successful or not. To further guarantee this, the checksum
file is now opened for writing in data sync mode (O_DSYNC) as well, such
that every write is a flush.