refactor(engine): re-add sent time to persistent state methods
This is a reluctant change. Initially I was hoping to find a way to update the sent time
when the command is actually being sent, instead of using the time when the persistent
update is made. However, updating the command sent time at the point when the command is
sent is a bigger undertaking. Commands can be sent by different classes, none of which
have the transient state immediately available.