fix(test-util): add condition for job timeouts
In our randomized property based tests we randomly decide, whether we
want to fail, complete or timeout an job. Timer
boundary events on service tasks and testing timeouts can make tests flaky, since
we adjust the clock which is consumed by both. It is not deterministic
which happens.

In order to fix that we add with this commit a condition that we not
trigger job timeouts if we already have timer boundary event.