fix(atomix): fixes memory leak in messaging service
Fixes a memory leak in the messaging service where connections opened
with keep alive created a long living opened future which was never
removed if the connection/channel initially failed to connect.

This was tested manually as there isn't an easy way to test for an
absence of memory leak; please refer to the associated PR for more.