fix(engine): use local vars for msg correlation
Variables could be overridden when a subprocess is completed,
specifically when non-interrupting events have happened on that
subprocess that placed temporary variables on it.

The biggest problem here is/was that the temporary variables are not
cleaned up until the element is completed. On completing the element,
the output mappings are applied and these then use all temporary
variables that still exist on it.

We can avoid this by reducing our temporary variables usage (something
we anyways want to do).

Message correlations with variables already create the variables
locally, so temporary variables are no longer needed for boundary
events.