fix(engine): remove event scope from previous deployment
Previous, on deploying a new version of a process, with a timer start event,
the timer was canceled and recreated, if the newer version had a timer start event.

The problem with this was that the event scope was created but never
deleted for that timer. With this commit on the ProcessCreatedApplier the eventScope for the previous deployment
is removed.